added class controller and post controller
